moves calibration configuration into a separate dialog
