+++ inlinino/resources/widget_hypernav_calibrate.ui
@@ -6,8 +6,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>297</width>
-    <height>320</height>
+    <width>287</width>
+    <height>193</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -32,52 +32,17 @@
       <string>Calibrate</string>
      </property>
      <layout class="QGridLayout" name="gridLayout">
-      <item row="3" column="1" alignment="Qt::AlignRight">
-       <widget class="QLabel" name="filename_lamp_label">
-        <property name="text">
-         <string>TextLabel</string>
-        </property>
-        <property name="wordWrap">
-         <bool>true</bool>
-        </property>
-       </widget>
-      </item>
-      <item row="3" column="2" alignment="Qt::AlignRight">
-       <widget class="QPushButton" name="browse_lamp_button">
-        <property name="text">
-         <string>Browse</string>
-        </property>
-       </widget>
-      </item>
-      <item row="4" column="0" alignment="Qt::AlignLeft">
-       <widget class="QLabel" name="label_3">
-        <property name="text">
-         <string>Plaque:</string>
-        </property>
-       </widget>
-      </item>
-      <item row="1" column="0" colspan="3">
+      <item row="1" column="0" colspan="2">
        <widget class="QLabel" name="label">
         <property name="text">
          <string>Select File to Analyze</string>
         </property>
        </widget>
       </item>
-      <item row="3" column="0" alignment="Qt::AlignLeft">
-       <widget class="QLabel" name="label_4">
-        <property name="text">
-         <string>Lamp:</string>
-        </property>
-       </widget>
+      <item row="2" column="0" colspan="3">
+       <widget class="QComboBox" name="filename_combobox"/>
       </item>
-      <item row="6" column="0" alignment="Qt::AlignLeft">
-       <widget class="QLabel" name="label_5">
-        <property name="text">
-         <string>Wavelength:</string>
-        </property>
-       </widget>
-      </item>
-      <item row="7" column="0" colspan="4">
+      <item row="4" column="0" colspan="3">
        <layout class="QHBoxLayout" name="horizontalLayout">
         <property name="topMargin">
          <number>0</number>
@@ -122,43 +87,6 @@
          <bool>true</bool>
         </property>
        </widget>
-      </item>
-      <item row="6" column="1" alignment="Qt::AlignRight">
-       <widget class="QLabel" name="filename_wavelength_label">
-        <property name="text">
-         <string>TextLabel</string>
-        </property>
-        <property name="wordWrap">
-         <bool>true</bool>
-        </property>
-       </widget>
-      </item>
-      <item row="4" column="1" alignment="Qt::AlignRight">
-       <widget class="QLabel" name="filename_plaque_label">
-        <property name="text">
-         <string>TextLabel</string>
-        </property>
-        <property name="wordWrap">
-         <bool>true</bool>
-        </property>
-       </widget>
-      </item>
-      <item row="6" column="2" alignment="Qt::AlignRight">
-       <widget class="QPushButton" name="browse_wavelength_button">
-        <property name="text">
-         <string>Browse</string>
-        </property>
-       </widget>
-      </item>
-      <item row="4" column="2" alignment="Qt::AlignRight">
-       <widget class="QPushButton" name="browse_plaque_button">
-        <property name="text">
-         <string>Browse</string>
-        </property>
-       </widget>
-      </item>
-      <item row="2" column="0" colspan="3">
-       <widget class="QComboBox" name="filename_combobox"/>
       </item>
      </layout>
     </widget>
