+++ inlinino/widgets/shared/file_label.py
@@ -4,7 +4,7 @@ class FileLabel():
     def __init__(self, qt_label: QtGui.QLabel):
         self.qt_label = qt_label
         self.absolute_file_path = None
-        self.qt_label.setText('')
+        self.qt_label.setText('...')
 
     def set_file(self, absolute_file_path):
         self.absolute_file_path = absolute_file_path
